azure-blob-download-using-mi/requirements.txt
azure-storage-blob
azure.identity